+++ app.json
@@ -35,7 +35,7 @@
     ],
     "extra": {
       "apiUrl": {
-        "development": "http://172.16.99.108:3000",
+        "development": "http://192.168.0.8:3000",
         "production": "https://your-production-api.com"
       }
     }
