+++ src/components/navigation/BottomTabNavigator.js
@@ -7,7 +7,6 @@ import AlbumScreen from "../../screens/album/AlbumScreen";
 import MainFeedScreen from "../../screens/MainFeedScreen";
 import MyPageScreen from "../../screens/MyPageScreen";
 
-import FriendsStackNavigator from "./FriendsStackNavigator";
 import CustomBottomTabBar from "./CustomBottomTabBar";
 import FriendsScreen from "../../screens/FriendsScreen";
